@@ -10,6 +10,9 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 import com.google.firebase.auth.FirebaseAuth;
+import com.squareup.picasso.Picasso;
+
+import java.io.File;
 import java.util.ArrayList;
 import endless.syria.sychat.R;
 import endless.syria.sychat.Utils.Models.CircleImageView;
@@ -18,6 +21,8 @@ import endless.syria.sychat.Utils.Models.MessageText;
 
 public class ChatAdapter extends RecyclerView.Adapter {
 
+    private static final int TYPE_ME = 0;
+    private final static int USER_TYPE = 1;
     private final ArrayList<MessageText> arrayList;
 
     private final String my = FirebaseAuth.getInstance().getCurrentUser().getDisplayName();
@@ -29,10 +34,11 @@ public class ChatAdapter extends RecyclerView.Adapter {
     @NonNull
     @Override
     public RecyclerView.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        if (viewType == 0){
-            return new MyMessageHolder(LayoutInflater.from(parent.getContext()).inflate(R.layout.mymessage_adapter,parent,false));
-        }else {
-            return new UserMessageHolder(LayoutInflater.from(parent.getContext()).inflate(R.layout.usermessage_adapter,parent,false));
+        if (viewType == TYPE_ME) {
+            return new MyMessageHolder(LayoutInflater.from(parent.getContext()).inflate(R.layout.mymessage_adapter, parent, false));
+        }
+        else {
+            return new UserMessageHolder(LayoutInflater.from(parent.getContext()).inflate(R.layout.usermessage_adapter, parent, false));
         }
     }
 
@@ -40,10 +46,10 @@ public class ChatAdapter extends RecyclerView.Adapter {
     public void onBindViewHolder(@NonNull RecyclerView.ViewHolder holder, int position) {
         MessageText messages = arrayList.get(position);
         switch (holder.getItemViewType()){
-            case 0 :
+            case TYPE_ME :
                 ((MyMessageHolder) holder).bind(messages);
                 break;
-            case 1 :
+            case USER_TYPE:
                 ((UserMessageHolder) holder).bind(messages);
                 break;
             default:
@@ -86,7 +92,7 @@ public class ChatAdapter extends RecyclerView.Adapter {
            // name.setText(messages.getUser());
             body.setText(messages.getMessage());
             time.setText(messages.getTime());
-            imageView.setImageURI(Uri.parse("sdcard/iData/users/" + messages.getUser() + "/" + messages.getUser() + ".jpg"));
+            Picasso.get().load(Uri.fromFile(new File("sdcard/iData/users/" + messages.getUser() + "/" + messages.getUser() + ".jpg"))).into(imageView);
         }
     }
 }
